refactor RoomChat
add abstracted network layer and repository layer
