+++ Linker/Modules/RoomChat/View/RoomChatViewController.swift
@@ -6,7 +6,6 @@
 //
 
 import UIKit
-//import Firebase
 
 class RoomChatViewController: UIViewController {
