+++ Linker/Services/Repository/Repository.swift
@@ -13,30 +13,4 @@ class Repository: RepositoryProtocol {
   init(networkManager: Networkable) {
     self.networkManager = networkManager
   }
-
-  func SignUp(withUsername username: String, email: String, password: String, compeletion: @escaping (Error?) -> Void) {
-    networkManager.SignUp(withUsername: username, email: email, password: password, compeletion: compeletion)
-  }
-
-  func signIn(withEmail email: String, password: String, compeletion: @escaping (SignInError?) -> Void) {
-    networkManager.signIn(withEmail: email, password: password, compeletion: compeletion)
-  }
-}
-
-extension Repository {
-  func createNewRoom(withName roomName: String, completion: @escaping VoidBlock) {
-    networkManager.createNewRoom(withName: roomName, completion: completion)
-  }
-
-  func observeRooms(completion: @escaping (Room) -> Void) {
-    networkManager.observeRooms(completion: completion)
-  }
-
-  func signOut(completion: @escaping VoidBlock) {
-    networkManager.signOut(completion: completion)
-  }
-
-  func isThereCurentUser() -> Bool {
-    networkManager.isThereCurentUser()
-  }
 }
